Restore functionality.
This addresses the following issues:

1. Body was not written if there was no data node. This restores the writing of
generated body of hex strings.

2. SIGPIPE on SSL_write due to the peer closing the connection would result in
the process abrubptly aborting. This gracefully handles the signal by reporting
an error and continuing processing the transactions.

3. The replay server would send body content in response to a HEAD request
while the client (correctly) did not expect body content. This fixes the
replay-server to not send body in response to a HEAD request.

4. The --no-proxy argument was not parsed correctly.

This also updates the build system to build with debug symbols with a release
build and updates the .gitignore for some scons build artifacts. It also tweaks
some of the status messages.